backend/requirements.txt
asgiref==3.7.2
astroid==2.15.8
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.7.0
bs4==0.0.1
cdifflib==1.2.6
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.7
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
decorator==4.4.2
dill==0.3.7
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-multiupload==0.6.1
django-tenant-schemas==1.11.0
django-tenants==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.4.2
drf-spectacular==0.26.5
eventlet==0.33.3
Flask==3.0.0
Flask-SocketIO==5.3.6
fuzzywuzzy==0.18.0
greenlet==3.0.0
gunicorn==21.2.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
imageio==2.31.3
imageio-ffmpeg==0.4.9
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
lazy-object-proxy==1.9.0
Markdown==3.4.4
MarkupSafe==2.1.3
mccabe==0.7.0
moviepy==1.0.3
numpy==1.26.3
ordered-set==4.1.0
packaging==23.1
Pillow==9.5.0
platformdirs==3.10.0
pluggy==1.3.0
proglog==0.1.10
psycopg2==2.9.7
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pylint==2.17.7
pyOpenSSL==23.2.0
pyotp==2.9.0
pypng==0.20220715.0
pytest==7.4.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.9.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.32.0
reportlab==4.0.5
requests==2.31.0
rpds-py==0.13.2
service-identity==23.1.0
simple-websocket==1.0.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
subprocess.run==0.0.8
tomlkit==0.12.1
tqdm==4.66.1
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
websocket-client==1.6.3
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.15.0
wsproto==1.2.0
zope.interface==6.1
